backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==2.0.0
asyncpg==0.29.0
sentence-transformers==2.2.2
spacy==3.7.2
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
faker==20.1.0
pydantic==2.5.0
python-multipart==0.0.6
passlib==1.7.4
google-generativeai==0.3.2
firebase-admin==6.1.0